+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - New argument to `search.pages()` to sort pages alphabetically
 - Added new argument `--help` and aliases `-h` and `-V` to cli
 - Added `eta` plugin, to support `Eta` template engine
+- New function `helper` to assign global helpers that can be used in the templates
 
 ### Removed
 - Arguments `path` and `recursive` in `search.pages()`